+++ mod/user/fromACL.js
@@ -18,7 +18,7 @@ const bcrypt = require('../utils/bcrypt')
 
 const crypto = require('crypto')
 
-const env = require('../../mapp_env.js')
+const env = require('../utils/processEnv.js')
 
 const acl = require('./acl')
